Added FileStorageTests.

+++ Chrono.FileSystem/DataProviders/FileDataProvider.cs
@@ -1,7 +1,9 @@
 ﻿using Chrono.DataProviders;
 using System;
 using System.IO;
+using System.Linq;
 using Chrono.Common;
+using Chrono.Exceptions;
 
 namespace Chrono.FileSystem.DataProviders
 {
@@ -25,23 +27,58 @@ namespace Chrono.FileSystem.DataProviders
 
         public Session GetSession(string sessionId)
         {
+            var result = GetSessionSave(sessionId);
+
+            if (result.IsSuccessful)
+            {
+                return result.Value;
+            }
+
+            throw new SessionNotFoundException();
+        }
+
+        public FuncResult<Session> GetSessionSave(string sessionId)
+        {
             Contract.NotNull<ArgumentNullException>(sessionId);
 
             var filePath = GetPathForSession(sessionId);
-            var serializedValue = ReadFromFile(filePath);
-            var session = Deserialize<Session>(serializedValue);
-            return session;
+            var resultFromFile = ReadFromFile(filePath);
+            if (resultFromFile.IsSuccessful)
+            {
+                var session = Deserialize<Session>(resultFromFile.Value);
+                return session.AsFuncResult();
+            }
+
+            return FuncResult<Session>.Failed();
         }
 
         public Snapshot GetSnapshot(string sessionId, string snapshotId)
+        {
+            var result = GetSnapshotSave(sessionId, snapshotId);
+
+            if (result.IsSuccessful)
+            {
+                return result.Value;
+            }
+
+            throw new SnapshotNotFoundException();
+        }
+
+        public FuncResult<Snapshot> GetSnapshotSave(string sessionId, string snapshotId)
         {
             Contract.NotNull<ArgumentNullException>(sessionId);
             Contract.NotNull<ArgumentNullException>(snapshotId);
 
             var filePath = GetPathForSnapshot(sessionId, snapshotId);
-            var serializedValue = ReadFromFile(filePath);
-            var snapshot = Deserialize<Snapshot>(serializedValue);
-            return snapshot;
+            var resultFromFile = ReadFromFile(filePath);
+
+            if (resultFromFile.IsSuccessful)
+            {
+                var snapshot = Deserialize<Snapshot>(resultFromFile.Value);
+                return snapshot.AsFuncResult();
+            }
+            
+            return FuncResult<Snapshot>.Failed();
         }
 
         public void RemoveSession(string sessionId)
@@ -50,7 +87,12 @@ namespace Chrono.FileSystem.DataProviders
 
             var filePath = GetPathForSession(sessionId);
             var directoryPath = Path.GetDirectoryName(filePath);
-            Directory.Delete(directoryPath);
+            File.Delete(filePath);
+
+            if (!Directory.GetFiles(directoryPath).Any())
+            {
+                Directory.Delete(directoryPath);
+            }
         }
 
         public void RemoveSnapshot(string sessionId, string snapshotId)
@@ -59,7 +101,10 @@ namespace Chrono.FileSystem.DataProviders
             Contract.NotNull<ArgumentNullException>(snapshotId);
 
             var filePath = GetPathForSnapshot(sessionId, snapshotId);
-            File.Delete(filePath);
+            if (File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
         }
 
         protected abstract string Serialize(object value);
@@ -71,10 +116,15 @@ namespace Chrono.FileSystem.DataProviders
             File.WriteAllText(filePath, content);
         }
 
-        protected string ReadFromFile(string filePath)
+        protected FuncResult<string> ReadFromFile(string filePath)
         {
+            if (!File.Exists(filePath))
+            {
+                return FuncResult.Failed<string>();
+            }
+
             var content = File.ReadAllText(filePath);
-            return content;
+            return content.AsFuncResult();
         }
 
         private string GetRootDirectoryPath()
